refactor: personDetail page scss 마이그레이션
